requirements.txt
tensorflow
transformers
pandas
scikit-learn
pyyaml
tqdm
google-generativeai
tf-keras